# backend/app/api/photos.py
"""
Photos API endpoints for CRUD operations.
"""
from fastapi import APIRouter, Depends, HTTPException, status, Query
from fastapi.responses import Response
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select, desc
from pydantic import BaseModel
from typing import Optional, List
from datetime import datetime

from app.core.database import get_db
from app.api.auth import get_current_user, get_current_user_or_token
from app.models.user import User
from app.models.photo import Photo, PhotoFile
from app.services.b2_service import b2_service

router = APIRouter()
router = APIRouter()


@router.get("/{photo_id}/download")
async def download_photo(
    photo_id: str,
    token: Optional[str] = Query(None),
    current_user: User = Depends(get_current_user_or_token),
    db: AsyncSession = Depends(get_db)
):
    """
    Download a photo's full resolution version.
    Accepts authentication via Authorization header or token query param.
    """
    # Get photo metadata
    result = await db.execute(
        select(Photo).where(
            Photo.photo_id == photo_id,
            Photo.user_id == current_user.user_id,
            Photo.deleted_at == None
        )
    )
    photo = result.scalar_one_or_none()
    
    if not photo:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Photo not found"
        )
    
    # Construct B2 key
    b2_key = f"uploads/{current_user.user_id}/{photo.photo_id}/original/{photo.filename}"
    
    try:
        # Download file bytes from B2
        file_bytes = b2_service.download_file_bytes(b2_key)
        
        # Return as streaming response
        return Response(
            content=file_bytes,
            media_type=photo.mime_type,
            headers={
                "Content-Disposition": f'inline; filename="{photo.filename}"',
                "Cache-Control": "private, max-age=3600"
            }
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to download photo: {str(e)}"
        )


@router.get("/{photo_id}/thumbnail/{size}")
async def get_thumbnail(
    photo_id: str,
    size: int,
    token: Optional[str] = Query(None),
    current_user: User = Depends(get_current_user_or_token),
    db: AsyncSession = Depends(get_db)
):
    """Get photo thumbnail (returns original for now, accepts token via query param or header)."""
    # Get photo metadata
    result = await db.execute(
        select(Photo).where(
            Photo.photo_id == photo_id,
            Photo.user_id == current_user.user_id,
            Photo.deleted_at == None
        )
    )
    photo = result.scalar_one_or_none()
    
    if not photo:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Photo not found"
        )
    
    # Construct B2 key
    b2_key = f"uploads/{current_user.user_id}/{photo.photo_id}/original/{photo.filename}"
    
    try:
        # Download file bytes from B2
        file_bytes = b2_service.download_file_bytes(b2_key)
        
        # Return as response
        return Response(
            content=file_bytes,
            media_type=photo.mime_type,
            headers={
                "Content-Disposition": f'inline; filename="{photo.filename}"',
                "Cache-Control": "private, max-age=3600"
            }
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to load thumbnail: {str(e)}"
        )


# Pydantic schemas
class PhotoResponse(BaseModel):
    photo_id: str
    filename: str
    mime_type: str
    size_bytes: int
    taken_at: Optional[datetime]
    uploaded_at: datetime
    caption: Optional[str]
    favorite: bool
    archived: bool
    thumb_urls: dict
    
    class Config:
        from_attributes = True


class PhotoListResponse(BaseModel):
    photos: List[PhotoResponse]
    next_cursor: Optional[str]
    has_more: bool


class UpdatePhotoRequest(BaseModel):
    caption: Optional[str] = None
    favorite: Optional[bool] = None
    archived: Optional[bool] = None


@router.get("", response_model=PhotoListResponse)
async def list_photos(
    cursor: Optional[str] = Query(None, description="Pagination cursor"),
    limit: int = Query(50, ge=1, le=200),
    sort: str = Query("taken_desc", regex="^(created_desc|created_asc|taken_desc|taken_asc)$"),
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_db)
):
    """
    List user's photos in timeline view.
    
    Supports pagination with cursor and sorting options.
    """
    # Build query
    query = select(Photo).where(
        Photo.user_id == current_user.user_id,
        Photo.deleted_at == None
    )
    
    # Apply sorting
    if sort == "taken_desc":
        query = query.order_by(desc(Photo.taken_at))
    elif sort == "taken_asc":
        query = query.order_by(Photo.taken_at)
    elif sort == "created_desc":
        query = query.order_by(desc(Photo.uploaded_at))
    else:  # created_asc
        query = query.order_by(Photo.uploaded_at)
    
    # Apply cursor pagination if provided
    # TODO: Implement cursor-based pagination properly
    
    # Limit results
    query = query.limit(limit + 1)  # Fetch one extra to check has_more
    
    result = await db.execute(query)
    photos = result.scalars().all()
    
    has_more = len(photos) > limit
    if has_more:
        photos = photos[:limit]
    
    # Build response
    photo_responses = []
    for photo in photos:
        # Generate thumbnail URLs (presigned from B2)
        thumb_urls = {
            "thumb_256": f"/api/v1/photos/{photo.photo_id}/thumbnail/256",
            "thumb_512": f"/api/v1/photos/{photo.photo_id}/thumbnail/512",
            "thumb_1024": f"/api/v1/photos/{photo.photo_id}/thumbnail/1024",
        }
        
        photo_responses.append(PhotoResponse(
            photo_id=str(photo.photo_id),
            filename=photo.filename,
            mime_type=photo.mime_type,
            size_bytes=photo.size_bytes,
            taken_at=photo.taken_at,
            uploaded_at=photo.uploaded_at,
            caption=photo.caption,
            favorite=photo.favorite,
            archived=photo.archived,
            thumb_urls=thumb_urls
        ))
    
    return PhotoListResponse(
        photos=photo_responses,
        next_cursor=None,  # TODO: Generate cursor from last photo
        has_more=has_more
    )


@router.get("/{photo_id}", response_model=PhotoResponse)
async def get_photo(
    photo_id: str,
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_db)
):
    """Get photo metadata by ID."""
    result = await db.execute(
        select(Photo).where(
            Photo.photo_id == photo_id,
            Photo.user_id == current_user.user_id,
            Photo.deleted_at == None
        )
    )
    photo = result.scalar_one_or_none()
    
    if not photo:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Photo not found"
        )
    
    # Generate thumbnail URLs
    thumb_urls = {
        "thumb_256": f"/api/v1/photos/{photo.photo_id}/thumbnail/256",
        "thumb_512": f"/api/v1/photos/{photo.photo_id}/thumbnail/512",
        "thumb_1024": f"/api/v1/photos/{photo.photo_id}/thumbnail/1024",
    }
    
    return PhotoResponse(
        photo_id=str(photo.photo_id),
        filename=photo.filename,
        mime_type=photo.mime_type,
        size_bytes=photo.size_bytes,
        taken_at=photo.taken_at,
        uploaded_at=photo.uploaded_at,
        caption=photo.caption,
        favorite=photo.favorite,
        archived=photo.archived,
        thumb_urls=thumb_urls
    )


@router.patch("/{photo_id}", response_model=PhotoResponse)
async def update_photo(
    photo_id: str,
    request: UpdatePhotoRequest,
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_db)
):
    """Update photo metadata (caption, favorite, archived)."""
    result = await db.execute(
        select(Photo).where(
            Photo.photo_id == photo_id,
            Photo.user_id == current_user.user_id,
            Photo.deleted_at == None
        )
    )
    photo = result.scalar_one_or_none()
    
    if not photo:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Photo not found"
        )
    
    # Update fields
    if request.caption is not None:
        photo.caption = request.caption
    if request.favorite is not None:
        photo.favorite = request.favorite
    if request.archived is not None:
        photo.archived = request.archived
    
    photo.updated_at = datetime.utcnow()
    
    await db.commit()
    await db.refresh(photo)
    
    thumb_urls = {
        "thumb_256": f"/api/v1/photos/{photo.photo_id}/thumbnail/256",
        "thumb_512": f"/api/v1/photos/{photo.photo_id}/thumbnail/512",
        "thumb_1024": f"/api/v1/photos/{photo.photo_id}/thumbnail/1024",
    }
    
    return PhotoResponse(
        photo_id=str(photo.photo_id),
        filename=photo.filename,
        mime_type=photo.mime_type,
        size_bytes=photo.size_bytes,
        taken_at=photo.taken_at,
        uploaded_at=photo.uploaded_at,
        caption=photo.caption,
        favorite=photo.favorite,
        archived=photo.archived,
        thumb_urls=thumb_urls
    )




@router.patch("/{photo_id}/favorite", status_code=status.HTTP_200_OK)
async def toggle_favorite(
    photo_id: str,
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_db)
):
    """Toggle favorite status of a photo."""
    result = await db.execute(
        select(Photo).where(
            Photo.photo_id == photo_id,
            Photo.user_id == current_user.user_id,
            Photo.deleted_at == None
        )
    )
    photo = result.scalar_one_or_none()
    
    if not photo:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Photo not found"
        )
    
    # Toggle favorite
    photo.favorite = not photo.favorite
    await db.commit()
    
    return {"favorite": photo.favorite}


@router.delete("/{photo_id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_photo(
    photo_id: str,
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_db)
):
    """Soft-delete a photo (moves to trash, 30-day retention) and delete from B2."""
    result = await db.execute(
        select(Photo).where(
            Photo.photo_id == photo_id,
            Photo.user_id == current_user.user_id,
            Photo.deleted_at == None
        )
    )
    photo = result.scalar_one_or_none()
    
    if not photo:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Photo not found"
        )
    
    # Delete from B2 first
    b2_key = f"uploads/{current_user.user_id}/{photo.photo_id}/original/{photo.filename}"
    try:
        # Get file info to get file_id for deletion
        import requests
        auth_token = b2_service.info.get_account_auth_token()
        download_url = f"{b2_service.info.get_download_url()}/file/{b2_service.get_bucket().name}/{b2_key}"
        
        # Try to get file info
        response = requests.head(download_url, headers={"Authorization": auth_token})
        if response.status_code == 200:
            # File exists, delete it using B2 API
            file_id = response.headers.get('x-bz-file-id')
            if file_id:
                b2_service.delete_file(file_id, b2_key)
    except Exception as e:
        # Log error but don't fail the delete - file might already be gone
        print(f"Warning: Failed to delete file from B2: {str(e)}")
    
    # Soft delete in database
    photo.deleted_at = datetime.utcnow()
    
    # Update user storage quota
    current_user.storage_used_bytes -= photo.size_bytes
    
    await db.commit()
    
    return None

@router.get("/{photo_id}/download")
async def download_photo(
    photo_id: str,
    token: Optional[str] = Query(None),
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_db)
):
    """Download original photo file from B2."""
    from fastapi.responses import StreamingResponse
    import io
    
    result = await db.execute(
        select(Photo).where(
            Photo.photo_id == photo_id,
            Photo.user_id == current_user.user_id,
            Photo.deleted_at == None
        )
    )
    photo = result.scalar_one_or_none()
    
    if not photo:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Photo not found"
        )
    
    # Generate B2 key
    b2_key = f"uploads/{current_user.user_id}/{photo.photo_id}/original/{photo.filename}"
    
    try:
        # Download file bytes from B2
        file_bytes = b2_service.download_file_bytes(b2_key)
        
        # Return as streaming response
        return StreamingResponse(
            io.BytesIO(file_bytes),
            media_type=photo.mime_type,
            headers={
                "Content-Disposition": f'inline; filename="{photo.filename}"'
            }
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=f"Photo file not found in storage: {str(e)}"
        )


@router.get("/{photo_id}/thumbnail/{size}")
async def get_thumbnail(
    photo_id: str,
    size: int,
    token: Optional[str] = Query(None),
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_db)
):
    """Get photo thumbnail (redirects to download for now, worker will generate thumbnails later)."""
    # For now, redirect to original download
    # Worker will later generate actual thumbnails
    return await download_photo(photo_id, current_user, db)