# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PhotoBomb API
  version: 1.0.0
  description: |
    Production-ready photo service API supporting uploads, metadata management, 
    search, albums, and sharing with presigned URLs and async processing.
  contact:
    name: PhotoBomb Team
    email: api@photobomb.example.com
servers:
  - url: https://api.photobomb.example.com/api/v1
    description: Production
  - url: https://api-staging.photobomb.example.com/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /photos:
    get:
      tags: [photos]
      summary: List user's photos (timeline)
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor (base64-encoded timestamp)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_desc, created_asc, taken_desc, taken_asc]
            default: taken_desc
      responses:
        '200':
          description: Photo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  photos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Photo'
                  next_cursor:
                    type: string
                    nullable: true
                  has_more:
                    type: boolean
  /photos/{photo_id}:
    get:
      tags: [photos]
      summary: Get photo metadata
      parameters:
        - name: photo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Photo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoDetailed'
        '404':
          description: Photo not found

    patch:
      tags: [photos]
      summary: Update photo metadata
      parameters:
        - name: photo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                caption:
                  type: string
                  maxLength: 2000
                favorite:
                  type: boolean
                archived:
                  type: boolean
      responses:
        '200':
          description: Photo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'

    delete:
      tags: [photos]
      summary: Soft-delete photo
      parameters:
        - name: photo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Photo deleted (moved to trash)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
  schemas:
...
    Photo:
      type: object
      properties:
        photo_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        filename:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        sha256:
          type: string
        taken_at:
          type: string
          format: date-time
          nullable: true
          description: From EXIF DateTimeOriginal
        uploaded_at:
          type: string
          format: date-time
        thumb_urls:
          type: object
          properties:
            thumb_256:
              type: string
              format: uri
            thumb_512:
              type: string
              format: uri
            thumb_1024:
              type: string
              format: uri
        caption:
          type: string
          nullable: true
        favorite:
          type: boolean
        archived:
          type: boolean
    PhotoDetailed:
      allOf:
        - $ref: '#/components/schemas/Photo'
        - type: object
          properties:
            exif:
              type: object
              properties:
                camera_make:
                  type: string
                camera_model:
                  type: string
                lens:
                  type: string
                iso:
                  type: integer
                aperture:
                  type: string
                shutter_speed:
                  type: string
                focal_length:
                  type: string
                gps_lat:
                  type: number
                  nullable: true
                gps_lng:
                  type: number
                  nullable: true
            original_url:
              type: string
              format: uri
              description: Signed download URL (5min expiry)
            faces:
              type: array
              items:
                type: object
                properties:
                  face_id:
                    type: string
                    format: uuid
                  bounding_box:
                    type: object
                    properties:
                      x:
                        type: number
                      y:
                        type: number
                      width:
                        type: number
                      height:
                        type: number
...